Rename RegressionModel into SKLearnModel
