+++ README.md
@@ -251,7 +251,7 @@ on bringing more models and features.
 | [GlobalNaiveDrift](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.global_baseline_models.html#darts.models.forecasting.global_baseline_models.GlobalNaiveDrift)                                                                                                   |                                                                                                                                                                                                                                   | ✅ ✅                                                          | 🔴 🔴 🔴                                                                 | 🔴 🔴                                                                    | ✅                                         |
 | [GlobalNaiveSeasonal](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.global_baseline_models.html#darts.models.forecasting.global_baseline_models.GlobalNaiveSeasonal)                                                                                             |                                                                                                                                                                                                                                   | ✅ ✅                                                          | 🔴 🔴 🔴                                                                 | 🔴 🔴                                                                    | ✅                                         |
 | **Regression Models**<br/>([GlobalForecastingModel](https://unit8co.github.io/darts/userguide/covariates.html#global-forecasting-models-gfms))                                                                                                                                            |                                                                                                                                                                                                                                   |                                                              |                                                                          |                                                                          |                                           |
-| [RegressionModel](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.regression_model.html#darts.models.forecasting.regression_model.RegressionModel): generic wrapper around any sklearn regression model                                                            |                                                                                                                                                                                                                                   | ✅ ✅                                                          | ✅ ✅ ✅                                                                    | 🔴 🔴                                                                    | ✅                                         |
+| [SKLearnModel](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.sklearn_model.html#darts.models.forecasting.sklearn_model.SKLearnModel): generic wrapper around any sklearn regression model                                                            |                                                                                                                                                                                                                                   | ✅ ✅                                                          | ✅ ✅ ✅                                                                    | 🔴 🔴                                                                    | ✅                                         |
 | [LinearRegressionModel](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.linear_regression_model.html#darts.models.forecasting.linear_regression_model.LinearRegressionModel)                                                                                       |                                                                                                                                                                                                                                   | ✅ ✅                                                          | ✅ ✅ ✅                                                                    | ✅ ✅                                                                      | ✅                                         |
 | [RandomForest](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.random_forest.html#darts.models.forecasting.random_forest.RandomForest)                                                                                                                             |                                                                                                                                                                                                                                   | ✅ ✅                                                          | ✅ ✅ ✅                                                                    | 🔴 🔴                                                                    | ✅                                         |
 | [LightGBMModel](https://unit8co.github.io/darts/generated_api/darts.models.forecasting.lgbm.html#darts.models.forecasting.lgbm.LightGBMModel)                                                                                                                                             |                                                                                                                                                                                                                                   | ✅ ✅                                                          | ✅ ✅ ✅                                                                    | ✅ ✅                                                                      | ✅                                         |
