+++ darts/models/forecasting/lgbm.py
@@ -16,10 +16,10 @@ from typing import Optional, Union
 import lightgbm as lgb
 
 from darts.logging import get_logger
-from darts.models.forecasting.regression_model import (
+from darts.models.forecasting.sklearn_model import (
     FUTURE_LAGS_TYPE,
     LAGS_TYPE,
-    RegressionModelWithCategoricalCovariates,
+    SKLearnModelWithCategoricalCovariates,
     _QuantileModelContainer,
 )
 from darts.timeseries import TimeSeries
@@ -32,7 +32,7 @@ from darts.utils.likelihood_models.sklearn import (
 logger = get_logger(__name__)
 
 
-class LightGBMModel(RegressionModelWithCategoricalCovariates):
+class LightGBMModel(SKLearnModelWithCategoricalCovariates):
     def __init__(
         self,
         lags: Optional[LAGS_TYPE] = None,
