+++ darts/models/forecasting/random_forest.py
@@ -20,16 +20,16 @@ from typing import Optional
 from sklearn.ensemble import RandomForestRegressor
 
 from darts.logging import get_logger
-from darts.models.forecasting.regression_model import (
+from darts.models.forecasting.sklearn_model import (
     FUTURE_LAGS_TYPE,
     LAGS_TYPE,
-    RegressionModel,
+    SKLearnModel,
 )
 
 logger = get_logger(__name__)
 
 
-class RandomForest(RegressionModel):
+class RandomForest(SKLearnModel):
     def __init__(
         self,
         lags: Optional[LAGS_TYPE] = None,
